Refactor ResponseHeaderDoneCallback out of GatewayBridge
